machineacs/pyproject.toml
[project]
name = "machineacs"
version = "0.9.0"
description = "A production-grade text and JSON cleaning pipeline."
dependencies = [
    "aiofiles>=25.1.0",
    "asyncio>=4.0.0",
    "fastapi>=0.128.0",
    "ijson",
    "pre-commit>=4.5.1",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.3.2",
    "pyspellchecker",
    "python-multipart>=0.0.21",
    "ruff>=0.14.10",
    "uvicorn>=0.40.0",
]

[project.scripts]
acs = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]



[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP", "SIM"]

ignore = []

[tool.ruff.lint.isort]
known-first-party = ["machineacs"]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
]
